Implementing the changes related to User Resorce

+++ src/main/java/com/revolut/assessment/dao/AcoountDao.java
@@ -1,12 +1,10 @@
 package com.revolut.assessment.dao;
 
-import java.math.BigDecimal;
-
 import com.revolut.assessment.model.Account;
 
 public interface AcoountDao {
 
-  void create(String number, BigDecimal balance, String currency, boolean status, BigDecimal limit);
+  void create(String number, Long balance, String currency, boolean status, Long limit);
 
   Account findAccountByNumber(String number);
