+++ src/main/java/com/revolut/assessment/services/AccountService.java
@@ -1,15 +1,13 @@
 package com.revolut.assessment.services;
 
-import java.math.BigDecimal;
-
 import com.revolut.assessment.model.Account;
 
 public interface AccountService {
 
-  void create(String number, BigDecimal balance, String currency, boolean status, BigDecimal limit);
+  void create(String number, Long balance, String currency, boolean status, Long limit);
 
   Account find(String number);
 
-  void transfer(String from, String to, String currency, BigDecimal amount);
+  void transfer(String from, String to, String currency, Long amount);
 
 }
